+++ lib/slack/web/api/endpoints/channels.rb
@@ -42,6 +42,8 @@ module Slack
           #   Start of time range of messages to include in results.
           # @option options [Object] :inclusive
           #   Include messages with latest or oldest timestamp in results.
+          # @option options [Object] :unreads
+          #   Include unread_count_display in the output?.
           # @see https://api.slack.com/methods/channels.history
           # @see https://github.com/dblock/slack-api-ref/blob/master/methods/channels.history.json
           def channels_history(options = {})